Fix logout. Split webapi into multiple apis.
