+++ actions/SessionActions.js
@@ -1,11 +1,11 @@
-import webapi from '../webapi/';
+import {userapi} from '../webapi/';
 
 var login = function(actionContext, payload, done) {
-    webapi.login(actionContext, payload, done);
+    userapi.login(actionContext, payload, done);
 };
 
 var logout = function(actionContext, payload, done) {
-    webapi.logout(actionContext, payload, done);
+    userapi.logout(actionContext, payload, done);
 };
 
 export {login, logout};
