+++ actions/UserActions.js
@@ -1,7 +1,7 @@
-import webapi from '../webapi/';
+import {userapi} from '../webapi/';
 
 var register = function(actionContext, payload, done) {
-    webapi.register(actionContext, payload, done);
+    userapi.register(actionContext, payload, done);
 };
 
 export {register};
