+++ fluxible-plugin-responses/index.js
@@ -13,21 +13,34 @@ export default function() {
                 plugActionContext: function (actionContext) {
                     actionContext.respondOk = function (message, redirectUrl, data) {
                         if (req.wantsJson()) {
-                            let ret = {success: true, message: message};
+                            let ret = {success: true};
                             if (data) {
                                 ret.data = data;
                             }
+                            if (message) {
+                                ret.message = message;
+                            }
+
                             res.status(200).send(ret);
                         } else {
-                            req.flash('success', message);
+                            if (message) {
+                                req.flash('success', message);
+                            }
                             res.redirect(redirectUrl);
                         }
                     };
                     actionContext.respondError = function (errorCode, message, redirectUrl) {
                         if (req.wantsJson()) {
-                            res.status(errorCode).send({success: false, message: message});
+                            let ret = {success: false};
+                            if (message) {
+                                ret.message = message;
+                            }
+
+                            res.status(errorCode).send(ret);
                         } else {
-                            req.flash('error', message);
+                            if (message) {
+                                req.flash('error', message);
+                            }
                             res.redirect(redirectUrl);
                         }
                     };
